backend/pyproject.toml
[project]
name = "gcp-vm-control"
version = "0.1.0"
description = "GCP-VM-control helps check the availability of already existing GCP VMs and turn them on and off as required by the user"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "google-api-python-client>=2.159.0",
    "google-auth>=2.37.0",
    "gdown>=4.6.1",
    "pydantic>=2.10.4",
    "sqlalchemy>=2.0.38",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.1"
]
